Clean up linting errors
