+++ contracts/cw-proposal-multiple/src/testing/mod.rs
@@ -1,4 +1,5 @@
-mod execute;
-mod instantiate;
-mod queries;
+pub mod do_votes;
+pub mod execute;
+pub mod instantiate;
+pub mod queries;
 mod tests;
